Consolidate schema and stats
