+++ src/schema.ts
@@ -1,8 +1,39 @@
 import {Type} from 'vega-lite/src/type';
+import {summary} from 'datalib/src/stats';
+
+import {EncodingQuery} from './query';
 
 export class Schema {
   private fieldSchemas: FieldSchema[];
   private fieldSchemaIndex: {[field: string]: FieldSchema};
+
+  /**
+   * Build a Schema object.
+   *
+   * @param data - a set of raw data
+   * @return a Schema object
+   */
+  public static build(data: any): Schema {
+    // create profiles for each variable
+    var summaries: Summary[] = summary(data);
+
+    var fieldSchemas: FieldSchema[] = summaries.map(function(summary) {
+      var field: string = summary.field;
+      var primitiveType: PrimitiveType = summary.type as any;
+      var type: Type = (primitiveType === PrimitiveType.NUMBER || primitiveType === PrimitiveType.INTEGER) ? Type.QUANTITATIVE:
+        primitiveType === PrimitiveType.DATE ? Type.TEMPORAL : Type.NOMINAL;
+      var distinct: number = summary.distinct;
+
+      return {
+        field: field,
+        type: type,
+        primitiveType: primitiveType,
+        distinct: distinct
+      };
+    });
+
+    return new Schema(fieldSchemas);
+  }
 
   constructor(fieldSchemas: FieldSchema[]) {
     this.fieldSchemas = fieldSchemas;
@@ -29,6 +60,18 @@ export class Schema {
   public type(field: string) {
     return this.fieldSchemaIndex[field] ? this.fieldSchemaIndex[field].type : null;
   }
+
+  public cardinality(encQ: EncodingQuery) {
+    if (encQ.aggregate || encQ.autoCount) {
+      return 1;
+    } else if (encQ.bin) {
+      return 1; // FIXME
+    } else if (encQ.timeUnit) {
+      return 1; // FIXME
+    }
+
+    return this.fieldSchemaIndex[encQ.field as string].distinct;
+  }
 }
 
 export enum PrimitiveType {
@@ -41,7 +84,7 @@ export enum PrimitiveType {
 
 export interface FieldSchema {
   field: string;
-  stats: FieldStats;
+  distinct: number;
   type?: Type;
   /** number, integer, string, date  */
   primitiveType: PrimitiveType;
